Build 2.0.0.16
---NEW *****Option in File menu to Import DDOBuilder V1 files*****
---Note that some work probably still needs to be done on this feature to get into all the nooks and crannies

---New bonus type "Untyped" added
---"Epic Mage Armor" and "Machrotechnic: Armor of Legends" now generate the correct effects
---Feat "Tactical master" now has the correct name
---Weapon breakdowns will now update on an enhancement train/revoke action
---"Machrotechnic: Armor of Legends" effects implemented when interacting with other Drive enhancements
---Machrotechnic Destiny Mantle now has the correct icon
---"Machrotechnic: Forcefield Generator" hit point bonus is now correctly a percentage increase
---"Ravager: Draconic Power Attack" (Dragon Lord version) now correctly costs 1ap/rank (Reported by Larendell)
---"Shiradi Champion: Fey Countenance"" now correctly requires a min spent of 20DPs to train (Reported by SpartanKiller13)
---"Tabaxi: Sharp Claws II" no longer crashes the software when selected
---Feats are now verified and revoked if required on a Race change
---Several enhancement tree items will now correctly show as a Clickie when selected
---All epic destiny cores that should grant 5 Ranged Power rather than 4 updated
---Old Epic Destiny fate Points awarded at Epic levels 28, 29 and 30 removed (Reported by CBDunk)
---Set bonuses will now be correctly applied when switching builds
---Paladins "Radiant Servant: Endless Turning" now correctly costs 1ap / rank (Reported by Warsaga)
---Missing "Deconstructor" augment added
---More item errors fixed (Down to 2600 errors)

+++ DDOBuilder/Quest.h
@@ -53,7 +53,7 @@ class Quest :
                 DL_FLAG(_, Elite) \
                 DL_FLAG(_, Reaper) \
                 DL_SIMPLE(_, int, Favor, 0) \
-                DL_VECTOR(_, int, Levels) \
+                DL_VECTOR(_, size_t, Levels) \
                 DL_FLAG(_, IgnoreForTotalFavor)
 
         DL_DECLARE_ACCESS(Quest_PROPERTIES)
